1. 优化 ComLog 的格式 2. 新增正则高亮设置框架 3. 命令批量运行是窗口置顶 4. 字体大小记忆 4. 各种选项记忆

+++ SuperCom/Config/AdvancedSendSettings.cs
@@ -20,8 +20,8 @@ namespace SuperCom.Config.WindowConfig
 
             return _instance;
         }
-        public int SideIndex { get; set; }
-        public int ComPortSelectedIndex { get; set; }
+        public long SideIndex { get; set; }
+        public long ComPortSelectedIndex { get; set; }
 
 
     }
